refactor: move erlang runtime
